feat: refactor state for search
