+++ App/AppRouter.swift
@@ -5,12 +5,6 @@ import SwiftUI
 import TodayUI
 
 public struct AppRouter: ViewModifier {
-    @Binding var searchText: String
-
-    public init(searchText: Binding<String>) {
-        _searchText = searchText
-    }
-
     public func body(content: Content) -> some View {
         content
             .navigationDestination(for: RouterDestination.self) { destination in
@@ -18,7 +12,7 @@ public struct AppRouter: ViewModifier {
                 case .kitchen:
                     KitchenView()
                 case .search:
-                    SearchView(searchText: $searchText)
+                    SearchView()
                 case .today:
                     TodayView()
                 }
@@ -27,7 +21,7 @@ public struct AppRouter: ViewModifier {
 }
 
 public extension View {
-    func withAppRouter(searchText: Binding<String>) -> some View {
-        modifier(AppRouter(searchText: searchText))
+    func withAppRouter() -> some View {
+        modifier(AppRouter())
     }
 }
