+++ App/AppTabsRoot.swift
@@ -16,9 +16,8 @@ struct AppTabRootView: View {
         TabView(selection: $router.selectedTab) {
             ForEach(AppTab.allCases) { tab in
                 NavigationStack(path: $router[tab]) {
-                    tab.rootView(searchText: $searchText)
+                    tab.rootView()
                         .environment(\.currentTab, tab)
-                        .navigationBarSearch(searchText: $searchText)
                 }
                 .tint(.white200)
                 .tabItem { tab.label }
@@ -31,77 +30,15 @@ struct AppTabRootView: View {
 @MainActor
 private extension AppTab {
     @ViewBuilder
-    func rootView(searchText: Binding<String>) -> some View {
+    func rootView() -> some View {
         switch self {
         case .today:
             TodayView()
         case .search:
-            SearchView(searchText: searchText)
+            SearchView()
         case .kitchen:
             KitchenView()
         }
-    }
-}
-
-public struct NavigatationBarSearch: ViewModifier {
-    @Binding var searchText: String
-    @Environment(Router.self) var router
-
-    public func body(content: Content) -> some View {
-        switch router.selectedTab {
-        case .search:
-            content.searchable(
-                text: $searchText,
-                placement: .navigationBarDrawer(displayMode: .always),
-                prompt: "What do you want to track next?"
-            )
-            .onAppear {
-                UISearchTextField.appearance().backgroundColor = .blue400
-                UISearchTextField.appearance().tintColor = .white200
-
-                UISearchTextField.appearance().borderStyle = .none
-                UISearchTextField.appearance().layer.cornerRadius = 10
-
-                UISearchTextField.appearance().attributedPlaceholder = NSAttributedString(
-                    string: "What do you want to track next?",
-                    attributes: [.foregroundColor: UIColor.gray200]
-                )
-
-                func searchBarImage() -> UIImage {
-                    let image = UIImage(systemName: "magnifyingglass")
-                    return image!.withTintColor(UIColor(.white200), renderingMode: .alwaysOriginal)
-                }
-
-                func clearButtonImage() -> UIImage {
-                    let image = UIImage(systemName: "xmark.circle.fill")
-                    return image!.withTintColor(UIColor(.blue800), renderingMode: .alwaysOriginal)
-                }
-
-                UISearchTextField.appearance(whenContainedInInstancesOf: [UISearchBar.self])
-                    .attributedPlaceholder = NSAttributedString(
-                        string: "What do you want to track next?",
-                        attributes: [.foregroundColor: UIColor(.white200)]
-                    )
-
-                UISearchBar.appearance(whenContainedInInstancesOf: [UINavigationBar.self]).setImage(
-                    searchBarImage(), for: .search, state: .normal
-                )
-                UISearchBar.appearance(whenContainedInInstancesOf: [UINavigationBar.self]).setImage(
-                    clearButtonImage(), for: .clear, state: .normal
-                )
-
-                UIBarButtonItem.appearance(whenContainedInInstancesOf: [UISearchBar.self])
-                    .setTitleTextAttributes([.foregroundColor: UIColor.white400], for: .normal)
-            }.foregroundColor(.white200)
-        case .today, .kitchen:
-            content
-        }
-    }
-}
-
-public extension View {
-    func navigationBarSearch(searchText: Binding<String>) -> some View {
-        modifier(NavigatationBarSearch(searchText: searchText))
     }
 }
