+++ App/KeepFresh.swift
@@ -1,7 +1,6 @@
 import DesignSystem
 import Router
 import SwiftUI
-import SwiftUIIntrospect
 
 public class FontRegistration {
     public static func registerFonts() {
