--- KeepFresh.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,15 +1,0 @@
-{
-  "originHash" : "7f661e1df466bbdb79cff179e6e921470fe87bae0ab0aed0f84344e7d0971ec1",
-  "pins" : [
-    {
-      "identity" : "swiftui-introspect",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/siteline/swiftui-introspect",
-      "state" : {
-        "revision" : "807f73ce09a9b9723f12385e592b4e0aaebd3336",
-        "version" : "1.3.0"
-      }
-    }
-  ],
-  "version" : 3
-}
