+++ Packages/Features/Sources/SearchUI/SearchView.swift
@@ -1,15 +1,77 @@
+import DesignSystem
+import Router
 import SwiftUI
 
 public struct SearchView: View {
-    public init(searchText: Binding<String>) {
-        _searchText = searchText
-    }
+    @State var searchText: String = ""
 
-    @Binding var searchText: String
+    public init() {}
 
     public var body: some View {
         VStack {
             Text("Recent Searches" + searchText).font(.headline).fontWeight(.bold).foregroundStyle(.black)
+        }.navigationBarSearch(searchText: $searchText)
+    }
+}
+
+public extension View {
+    func navigationBarSearch(searchText: Binding<String>) -> some View {
+        modifier(NavigatationBarSearch(searchText: searchText))
+    }
+}
+
+public struct NavigatationBarSearch: ViewModifier {
+    @Binding var searchText: String
+    @Environment(Router.self) var router
+
+    public func body(content: Content) -> some View {
+        switch router.selectedTab {
+        case .search:
+            content.searchable(
+                text: $searchText,
+                placement: .navigationBarDrawer(displayMode: .always),
+                prompt: "What do you want to track next?"
+            )
+            .onAppear {
+                UISearchTextField.appearance().backgroundColor = .blue400
+                UISearchTextField.appearance().tintColor = .white200
+
+                UISearchTextField.appearance().borderStyle = .none
+                UISearchTextField.appearance().layer.cornerRadius = 10
+
+                UISearchTextField.appearance().attributedPlaceholder = NSAttributedString(
+                    string: "What do you want to track next?",
+                    attributes: [.foregroundColor: UIColor.gray200]
+                )
+
+                func searchBarImage() -> UIImage {
+                    let image = UIImage(systemName: "magnifyingglass")
+                    return image!.withTintColor(UIColor(.white200), renderingMode: .alwaysOriginal)
+                }
+
+                func clearButtonImage() -> UIImage {
+                    let image = UIImage(systemName: "xmark.circle.fill")
+                    return image!.withTintColor(UIColor(.blue800), renderingMode: .alwaysOriginal)
+                }
+
+                UISearchTextField.appearance(whenContainedInInstancesOf: [UISearchBar.self])
+                    .attributedPlaceholder = NSAttributedString(
+                        string: "What do you want to track next?",
+                        attributes: [.foregroundColor: UIColor(.white200)]
+                    )
+
+                UISearchBar.appearance(whenContainedInInstancesOf: [UINavigationBar.self]).setImage(
+                    searchBarImage(), for: .search, state: .normal
+                )
+                UISearchBar.appearance(whenContainedInInstancesOf: [UINavigationBar.self]).setImage(
+                    clearButtonImage(), for: .clear, state: .normal
+                )
+
+                UIBarButtonItem.appearance(whenContainedInInstancesOf: [UISearchBar.self])
+                    .setTitleTextAttributes([.foregroundColor: UIColor.white400], for: .normal)
+            }.foregroundColor(.white200)
+        case .today, .kitchen:
+            content
         }
     }
 }
